refactor(meta): rename governance package -> meta + add migration standard
The package that defines what a package is is renamed governance -> meta (it owns
the model's base + governance extension + meta-index data — meta covers all three;
'governance' only named the extension). Mechanical: git mv common/governance ->
common/meta + rewrite common.governance -> common.meta across 33 files
(unambiguous; no collision with the separate SSOT-governance tooling), contract
name=meta. Adds common/meta/migration-standard.md (the target architecture: ~10
packages, base/extension/data, AC-<pkg>.<entity>.<seq>, FE via interface+vectors,
AC-as-migration-unit DoD).
