+++ apps/backend/src/counter/README.md
@@ -7,7 +7,7 @@ ubiquitous language, contract, roles, storage, and governance — lives in
 
 - Spec / ubiquitous language: [`common/counter/readme.md`](../../../../common/counter/readme.md)
 - Machine contract (interface, invariants, roadmap ACs): [`common/counter/contract.py`](../../../../common/counter/contract.py)
-- Package model: [`common/governance/readme.md`](../../../../common/governance/readme.md)
+- Package model: [`common/meta/readme.md`](../../../../common/meta/readme.md)
 
 The code here converges by role — `types/` (nouns + events), `ops/` (verbs over
 the store port), `store/` (the `CounterRepository` port + SQL adapter), `api/`
