+++ common/config/contract.py
@@ -12,7 +12,7 @@ A curated published-language surface is a future cleanup.
 
 from __future__ import annotations
 
-from common.governance.package_contract import Invariant, PackageContract
+from common.meta.package_contract import Invariant, PackageContract
 
 CONTRACT = PackageContract(
     name="config",
