+++ common/observability/contract.py
@@ -9,7 +9,7 @@ surface is a future cleanup.
 
 from __future__ import annotations
 
-from common.governance.package_contract import Invariant, PackageContract
+from common.meta.package_contract import Invariant, PackageContract
 
 CONTRACT = PackageContract(
     name="observability",
