+++ common/platform/contract.py
@@ -24,7 +24,7 @@ capability of the platform substrate); ``kernel`` is its honest DAG rank.
 
 from __future__ import annotations
 
-from common.governance.package_contract import (
+from common.meta.package_contract import (
     ACRecord,
     Invariant,
     PackageContract,
