+++ common/quantity/contract.py
@@ -12,7 +12,7 @@ table; moving that ownership into the roadmap is a tracked follow-up.
 
 from __future__ import annotations
 
-from common.governance.package_contract import Invariant, PackageContract
+from common.meta.package_contract import Invariant, PackageContract
 
 CONTRACT = PackageContract(
     name="quantity",
