+++ docs/project/traceability-exceptions.md
@@ -128,9 +128,9 @@ explicit AC IDs for the behavior.
 | `tests/tooling/test_brokerage_prompt_contract.py` | `docs/ssot/extraction.md` |
 | `tests/tooling/test_check_env_keys.py` | `docs/ssot/development.md` |
 | `tests/tooling/test_check_manifest.py` | `docs/ssot/MANIFEST.yaml` |
-| `tests/tooling/test_check_package_contract.py` | `common/governance/readme.md` |
+| `tests/tooling/test_check_package_contract.py` | `common/meta/readme.md` |
 | `tests/tooling/test_check_ssot_ownership.py` | `docs/ssot/MANIFEST.yaml` |
-| `tests/tooling/test_counter_package.py` | `common/governance/readme.md` |
+| `tests/tooling/test_counter_package.py` | `common/meta/readme.md` |
 | `tests/tooling/test_coverage_analyzer.py` | `docs/ssot/coverage.md` |
 | `tests/tooling/test_migration_safety_gates.py` | `docs/ssot/authority-tiers.md` |
 | `tests/tooling/test_delivery_gates_contract.py` | `docs/ssot/delivery-gates.yaml` |
