+++ docs/ssot/MANIFEST.yaml
@@ -78,15 +78,15 @@ concepts:
       - tests/tooling/test_quantity_api_parity.py
 
   package_model:
-    owner: common/governance/readme.md#package-model
-    description: A package = a DDD bounded context (readme.md + PackageContract + types/ops/store/api roles + __all__ published language); three classes core/platform/kernel; governance is computed from contracts by check_package_contract. The model self-hosts in the common/governance meta package.
+    owner: common/meta/readme.md#package-model
+    description: A package = a DDD bounded context (readme.md + PackageContract + types/ops/store/api roles + __all__ published language); three classes core/platform/kernel; governance is computed from contracts by check_package_contract. The model self-hosts in the common/meta package.
     family: platform
     kind: concept
     authority: documented_contract
     cross_refs:
-      - common/governance/package_contract.py
-      - common/governance/check_package_contract.py
-      - common/governance/contract.py
+      - common/meta/package_contract.py
+      - common/meta/check_package_contract.py
+      - common/meta/contract.py
       - common/counter/readme.md
       - common/counter/contract.py
       - apps/backend/src/counter/__init__.py
